move functions to util that do not perform file creation
