+++ hyp3_metadata/__init__.py
@@ -3,7 +3,7 @@
 # FIXME: Python 3.8+ this should be `from importlib.metadata...`
 from importlib_metadata import PackageNotFoundError, version
 
-from hyp3_metadata.create import create_metadata_file_set
+from hyp3_metadata.create import create_metadata_file_set_rtc
 
 try:
     __version__ = version(__name__)
@@ -14,4 +14,4 @@ except PackageNotFoundError:
           'Or, to just get the version number use:\n'
           '   python setup.py --version')
 
-__all__ = ['__version__', 'create_metadata_file_set']
+__all__ = ['__version__', 'create_metadata_file_set_rtc']
