+++ tests/test_rtc_gamma.py
@@ -2,12 +2,13 @@ from datetime import datetime
 
 import pytest
 
+import hyp3_metadata.util
 from hyp3_metadata import create
 from hyp3_metadata.util import SUPPORTED_DEMS
 
 
 def test_create_rtc_gamma_readme(product_dir):
-    payload = create.marshal_metadata(
+    payload = hyp3_metadata.util.marshal_metadata(
         product_dir=product_dir,
         granule_name='S1A_IW_SLC__1SSV_20150621T120220_20150621T120232_006471_008934_72D8',
         dem_name='SRTMGL1',
@@ -25,7 +26,7 @@ def test_create_rtc_gamma_readme(product_dir):
 
 
 def test_rtc_gamma_product(product_dir):
-    payload = create.marshal_metadata(
+    payload = hyp3_metadata.util.marshal_metadata(
         product_dir=product_dir,
         granule_name='S1A_IW_SLC__1SSV_20150621T120220_20150621T120232_006471_008934_72D8',
         dem_name='SRTMGL1',
@@ -48,7 +49,7 @@ def test_rtc_gamma_product(product_dir):
 
 def test_create_dem_xml(product_dir):
     for dem_name in SUPPORTED_DEMS:
-        payload = create.marshal_metadata(
+        payload = hyp3_metadata.util.marshal_metadata(
             product_dir=product_dir,
             granule_name='S1A_IW_SLC__1SSV_20150621T120220_20150621T120232_006471_008934_72D8',
             dem_name=dem_name,
@@ -79,7 +80,7 @@ def test_create_dem_xml(product_dir):
 
 
 def test_create_browse_xmls(product_dir):
-    payload = create.marshal_metadata(
+    payload = hyp3_metadata.util.marshal_metadata(
         product_dir=product_dir,
         granule_name='S1A_IW_SLC__1SSV_20150621T120220_20150621T120232_006471_008934_72D8',
         dem_name='SRTMGL1',
@@ -101,7 +102,7 @@ def test_create_browse_xmls(product_dir):
 
 
 def test_rtc_gamma_area(product_dir):
-    payload = create.marshal_metadata(
+    payload = hyp3_metadata.util.marshal_metadata(
         product_dir=product_dir,
         granule_name='S1A_IW_SLC__1SSV_20150621T120220_20150621T120232_006471_008934_72D8',
         dem_name='SRTMGL1',
@@ -119,7 +120,7 @@ def test_rtc_gamma_area(product_dir):
 
 
 def test_rtc_gamma_inc_map(product_dir):
-    payload = create.marshal_metadata(
+    payload = hyp3_metadata.util.marshal_metadata(
         product_dir=product_dir,
         granule_name='S1A_IW_SLC__1SSV_20150621T120220_20150621T120232_006471_008934_72D8',
         dem_name='SRTMGL1',
@@ -137,7 +138,7 @@ def test_rtc_gamma_inc_map(product_dir):
 
 
 def test_rtc_gamma_ls_map(product_dir):
-    payload = create.marshal_metadata(
+    payload = hyp3_metadata.util.marshal_metadata(
         product_dir=product_dir,
         granule_name='S1A_IW_SLC__1SSV_20150621T120220_20150621T120232_006471_008934_72D8',
         dem_name='SRTMGL1',
@@ -155,7 +156,7 @@ def test_rtc_gamma_ls_map(product_dir):
 
 
 def test_rtc_gamma_rgb(product_dir):
-    payload = create.marshal_metadata(
+    payload = hyp3_metadata.util.marshal_metadata(
         product_dir=product_dir,
         granule_name='S1A_IW_SLC__1SDV_20150621T120220_20150621T120232_006471_008934_72D8',
         dem_name='SRTMGL1',
@@ -173,7 +174,7 @@ def test_rtc_gamma_rgb(product_dir):
 
 
 def test_rtc_gamma_all_files(product_dir):
-    files = create.create_metadata_file_set(
+    files = create.create_metadata_file_set_rtc(
         product_dir=product_dir,
         granule_name='S1A_IW_SLC__1SSV_20150621T120220_20150621T120232_006471_008934_72D8',
         dem_name='SRTMGL1',
@@ -202,34 +203,34 @@ def test_rtc_gamma_all_files(product_dir):
 
 def test_thumbnail_no_such_reference_file(test_data_folder):
     reference_file = test_data_folder / 'no_such_file'
-    assert create.get_thumbnail_encoded_string(reference_file) == ''
+    assert hyp3_metadata.util.get_thumbnail_encoded_string(reference_file) == ''
 
 
 def test_thumbnail_reference_file_is_browse(test_data_folder):
     reference_file = test_data_folder / 'rtc.png'
-    encoded_string = create.get_thumbnail_encoded_string(reference_file)
+    encoded_string = hyp3_metadata.util.get_thumbnail_encoded_string(reference_file)
     assert len(encoded_string) == 844
 
 
 def test_thumbnail_reference_file_is_pol(test_data_folder):
     reference_file = test_data_folder / 'rtc_VV.png'
-    encoded_string = create.get_thumbnail_encoded_string(reference_file)
+    encoded_string = hyp3_metadata.util.get_thumbnail_encoded_string(reference_file)
     assert len(encoded_string) == 844
 
     reference_file = test_data_folder / 'rtc_VH.png'
-    encoded_string = create.get_thumbnail_encoded_string(reference_file)
+    encoded_string = hyp3_metadata.util.get_thumbnail_encoded_string(reference_file)
     assert len(encoded_string) == 844
 
 
 def test_thumbnail_reference_file_is_dem(test_data_folder):
     reference_file = test_data_folder / 'rtc_dem.tif'
-    encoded_string = create.get_thumbnail_encoded_string(reference_file)
+    encoded_string = hyp3_metadata.util.get_thumbnail_encoded_string(reference_file)
     assert len(encoded_string) == 844
 
 
 def test_decode_product():
     name = 'S1A_IW_20150621T120220_SVP_RTC10_G_sauned_F8E2'
-    assert create.decode_product(name) == {
+    assert hyp3_metadata.util.decode_product(name) == {
         'pixel_spacing': 10,
         'radiometry': 'sigma-0',
         'scale': 'amplitude',
@@ -241,7 +242,7 @@ def test_decode_product():
     }
 
     name = 'S1B_IW_20150621T120220_DHR_RTC30_G_gpwfcm_F8E2'
-    assert create.decode_product(name) == {
+    assert hyp3_metadata.util.decode_product(name) == {
         'pixel_spacing': 30,
         'radiometry': 'gamma-0',
         'scale': 'power',
